Enhance project functionality and style
