+++ index.html
@@ -9,8 +9,13 @@
   <link rel="stylesheet" href="src/styles.css" />
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css" />
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600;700&display=swap" rel="stylesheet">
+  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&display=swap" rel="stylesheet">
 </head>
 <body>
+
+  <div class="preloader">
+    <div class="loader"></div>
+  </div>
 
   <nav class="navbar">
     <div class="logo">AY</div>
@@ -98,6 +103,36 @@
         <div class="skill-item"><p>Bible-Based AI Pipelines</p></div>
         <div class="skill-item"><p>Multilingual UX + NLP</p></div>
         <div class="skill-item"><p>Ultra-Hard DSA & AI Problem Design</p></div>
+      </div>
+    </div>
+  </section>
+
+  <section id="services" class="services-section animate-on-scroll">
+    <div class="section-header">
+      <h2>My Services</h2>
+      <div class="underline"></div>
+    </div>
+    <div class="services-grid">
+      <div class="service-card">
+        <div class="service-icon">
+          <i class="fas fa-code"></i>
+        </div>
+        <h3>Web Development</h3>
+        <p>I build responsive and dynamic websites using the latest technologies.</p>
+      </div>
+      <div class="service-card">
+        <div class="service-icon">
+          <i class="fas fa-pen-nib"></i>
+        </div>
+        <h3>Content Writing</h3>
+        <p>I write engaging and SEO-friendly content for your website or blog.</p>
+      </div>
+      <div class="service-card">
+        <div class="service-icon">
+          <i class="fas fa-language"></i>
+        </div>
+        <h3>Translation</h3>
+        <p>I provide accurate and reliable translation services from English to Urdu and vice versa.</p>
       </div>
     </div>
   </section>
@@ -210,6 +245,7 @@
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/typed.js/2.0.12/typed.min.js"></script>
   <script type="module" src="src/script.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/vanilla-tilt/1.7.0/vanilla-tilt.min.js"></script>
   <div class="cursor"></div>
 </body>
 </html>
